kafka_receiver/requirements.txt
kafka-python==2.0.2
cryptography==41.0.7
